# services/ai-gateway/openapi/techflow-ai-gateway-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TechFlow AI Gateway",
    "description": "Issue #41 API, database, provider profile, and mock adapter foundation.",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources": {
      "post": {
        "summary": "Create Source",
        "operationId": "createSource",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources/{sourceId}": {
      "get": {
        "summary": "Get Source",
        "operationId": "getSource",
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sourceid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Withdraw Source",
        "operationId": "withdrawSource",
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sourceid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/compatibility-sets": {
      "post": {
        "summary": "Create Compatibility Set",
        "operationId": "createCompatibilitySet",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatibilitySetCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources/{sourceId}/approve": {
      "post": {
        "summary": "Approve Source",
        "operationId": "approveSource",
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sourceid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources/{sourceId}/ingestions": {
      "post": {
        "summary": "Create Ingestion",
        "operationId": "createIngestion",
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sourceid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{jobId}": {
      "get": {
        "summary": "Get Job",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jobid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rag/query": {
      "post": {
        "summary": "Query Rag",
        "operationId": "queryRag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evaluations/runs": {
      "post": {
        "summary": "Create Evaluation Run",
        "operationId": "createEvaluationRun",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evaluations/runs/{runId}": {
      "get": {
        "summary": "Get Evaluation Run",
        "operationId": "getEvaluationRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Runid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiMeta": {
        "properties": {
          "correlationId": {
            "type": "string",
            "title": "Correlationid"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          },
          "apiVersion": {
            "type": "string",
            "title": "Apiversion",
            "default": "v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "correlationId"
        ],
        "title": "ApiMeta"
      },
      "CompatibilityMember": {
        "properties": {
          "sourceVersionId": {
            "type": "string",
            "format": "uuid",
            "title": "Sourceversionid"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sourceVersionId"
        ],
        "title": "CompatibilityMember"
      },
      "CompatibilitySetCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "title": "Name"
          },
          "product": {
            "type": "string",
            "const": "ABLESTACK",
            "title": "Product",
            "default": "ABLESTACK"
          },
          "productVersion": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Productversion"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/CompatibilityMember"
            },
            "type": "array",
            "maxItems": 16,
            "minItems": 1,
            "title": "Members"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "productVersion",
          "members"
        ],
        "title": "CompatibilitySetCreateRequest"
      },
      "Envelope": {
        "properties": {
          "data": {
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope"
      },
      "EvaluationRunCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "title": "Name"
          },
          "sourceProfileIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "pattern": "^[A-Z][A-Z0-9_]{2,63}$"
                },
                "type": "array",
                "maxItems": 9,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceprofileids"
          },
          "compatibilitySetId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compatibilitysetid"
          },
          "providerProfileId": {
            "type": "string",
            "enum": [
              "OPENAI_RAG_DEFAULT_V1",
              "OPENAI_RAG_ESCALATION_V1"
            ],
            "title": "Providerprofileid"
          },
          "requestedBy": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_.:@-]{3,128}$",
            "title": "Requestedby"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "providerProfileId",
          "requestedBy"
        ],
        "title": "EvaluationRunCreateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestionCreateRequest": {
        "properties": {
          "requestedBy": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_.:@-]{3,128}$",
            "title": "Requestedby"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestedBy"
        ],
        "title": "IngestionCreateRequest"
      },
      "QueryRequest": {
        "properties": {
          "queryId": {
            "type": "string",
            "format": "uuid",
            "title": "Queryid"
          },
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Question"
          },
          "sourceProfileIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "pattern": "^[A-Z][A-Z0-9_]{2,63}$"
                },
                "type": "array",
                "maxItems": 9,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceprofileids"
          },
          "compatibilitySetId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compatibilitysetid"
          },
          "locale": {
            "type": "string",
            "enum": [
              "ko-KR",
              "en-US"
            ],
            "title": "Locale",
            "default": "ko-KR"
          },
          "classification": {
            "type": "string",
            "const": "D0",
            "title": "Classification",
            "default": "D0"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "queryId",
          "question"
        ],
        "title": "QueryRequest"
      },
      "SourceApprovalRequest": {
        "properties": {
          "approvedBy": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_.:@-]{3,128}$",
            "title": "Approvedby"
          },
          "decisionNote": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisionnote"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "approvedBy"
        ],
        "title": "SourceApprovalRequest"
      },
      "SourceCreateRequest": {
        "properties": {
          "sourceProfileId": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$",
            "title": "Sourceprofileid"
          },
          "repository": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$",
            "title": "Repository"
          },
          "branch": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Branch"
          },
          "commit": {
            "type": "string",
            "pattern": "^[0-9a-f]{40}$",
            "title": "Commit"
          },
          "sourceKind": {
            "$ref": "#/components/schemas/SourceKind"
          },
          "classification": {
            "type": "string",
            "const": "D0",
            "title": "Classification",
            "default": "D0"
          },
          "licenseSpdx": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensespdx"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sourceProfileId",
          "repository",
          "branch",
          "commit",
          "sourceKind"
        ],
        "title": "SourceCreateRequest"
      },
      "SourceKind": {
        "type": "string",
        "enum": [
          "DOCUMENTATION",
          "SOURCE_CODE"
        ],
        "title": "SourceKind"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}